+++ hypervisor/boot/multiboot/multiboot2.c
@@ -7,6 +7,7 @@
 #include <types.h>
 #include <errno.h>
 #include <boot.h>
+#include <multiboot.h>
 #include <pgtable.h>
 #include "multiboot_priv.h"
