+++ hypervisor/include/arch/x86/vm_config.h
@@ -9,8 +9,8 @@
 
 #include <types.h>
 #include <pci.h>
-#include <boot.h>
 #include <board_info.h>
+#include <multiboot.h>
 #include <acrn_common.h>
 #include <vm_uuids.h>
 #include <vm_configurations.h>
